Extracted interface from implementation for messenger
