+++ src/packet_handlers/initiation_protocols.rs
@@ -11,6 +11,6 @@ pub mod handshake;
 pub mod login;
 
 use super::game_state;
-use super::messenger;
+use super::interfaces;
 use super::packet;
 use super::translation;
